Refactored solve.
